pyproject.toml
[project]
name = "veinpay"
version = "0.1.0"
description = "Contactless Vein Pattern Recognition System for Secure and Seamless Payment Authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
    "streamlit-option-menu>=0.4.0",
    "tensorflow>=2.20.0",
    "uvicorn>=0.38.0",
]
